refactor(org-lens): move people tabs to @lfx-one/shared
LFXV2-1869. Address CodeRabbit review on PR #760:

- Promote PeopleTabConfig + PeopleTabId to
  packages/shared/src/interfaces/org-people.interface.ts.
- Promote PEOPLE_TABS, DEFAULT_PEOPLE_TAB_ID, and VALID_PEOPLE_TAB_IDS
  to packages/shared/src/constants/org-people.constants.ts and
  re-export from the constants barrel.
- Update OrgPeopleComponent to import the contract from
  @lfx-one/shared/constants and the types from
  @lfx-one/shared/interfaces. No behaviour change; the component
  drops its local interface, type, and module-level consts so the
  apps/lfx-one/ layer holds no shared contract per repo rule.
- Mark the tab icons aria-hidden so assistive tech skips the
  decorative <i> glyphs (CodeRabbit a11y nit).

Matches the marketing-impact precedent
(MARKETING_IMPACT_TABS / MarketingImpactTabOption already split
across constants/interfaces in @lfx-one/shared).
