+++ apps/lfx-one/src/app/modules/dashboards/org/org-people/org-people.component.html
@@ -26,7 +26,7 @@
         [class.text-gray-500]="activeTab() !== tab.id"
         [class.hover:text-gray-700]="activeTab() !== tab.id"
         (click)="switchTab(tab.id)">
-        <i [class]="tab.icon"></i>
+        <i [class]="tab.icon" aria-hidden="true"></i>
         {{ tab.label }}
       </button>
     }
