+++ packages/shared/src/constants/index.ts
@@ -54,4 +54,5 @@ export * from './org-lens.constants';
 export * from './feature-flags.constants';
 export * from './foundation-projects.constants';
 export * from './marketing-impact.constants';
+export * from './org-people.constants';
 export * from './individual-enrollment-catalog';
